+++ CHANGELOG.md
@@ -18,6 +18,7 @@ For a full diff see [`2022011100...main`](https://gitlab.matrix.org/new-vector/m
 - Started renaming Matrix chat rooms when a staff member updates a group ([#64](https://gitlab.matrix.org/new-vector/moodle-mod_matrix/-/merge_requests/64))
 - Renamed and moved `EventSubscriber` to allow moodle to auto-load cached observers ([#65](https://gitlab.matrix.org/new-vector/moodle-mod_matrix/-/merge_requests/65))
 - Started removing all rooms for a group when a staff member removes a group ([#66](https://gitlab.matrix.org/new-vector/moodle-mod_matrix/-/merge_requests/66))
+- Stopped using `json_encode()` for room URLs ([#68](https://gitlab.matrix.org/new-vector/moodle-mod_matrix/-/merge_requests/68))
 
 ## [`2022011101`](https://gitlab.matrix.org/new-vector/moodle-mod_matrix/-/tags/2022011101)